add __main__, add verbose mode toggle
